+++ CHANGES.rst
@@ -4,9 +4,12 @@ Products.CMFCore Changelog
 2.4.0b7 (unreleased)
 --------------------
 
+- Deprecated unused method ``PortalCatalogProcessor.get_dispatcher``
+  (`#57 <https://github.com/zopefoundation/Products.CMFCore/issues/57>`_)
+
 - Backport plone/collective.indexing#16 to use the catalog queue
   also when reindexing the object security
-  (`#58 <https://github.com/zopefoundation/Products.CMFCore/issues/58>_`)
+  (`#58 <https://github.com/zopefoundation/Products.CMFCore/issues/58>`_)
 
 - Code cleanup: avoid to call the ``keys`` method on dictionary
   when not needed